修复: 推理脚本dtype问题并完成首次测试
修复内容:
- 修复自定义层dtype不匹配问题（BFloat16 vs Float）
- 修复BFloat16转numpy的兼容性问题
- 确保推理时所有层使用正确的dtype

测试结果:
- ✅ 成功预测40残基蛋白质的距离矩阵
- ✅ 距离矩阵对称，对角线为0
- ✅ 距离范围: 0-42Å (合理)
- ✅ MDS重建相关系数: 0.706 (接近目标0.7)
- ✅ RMSE: 8.57Å (良好)
- ✅ PDB文件格式正确

输出文件:
- test_pred.pdb (3.3KB) - 3D坐标
- test_pred_distmat.npy (6.4KB) - 距离矩阵

下一步: 验证集批量评估

+++ infer_distmat.py
@@ -168,6 +168,10 @@ def load_model(model_path: str, device: str = 'cuda', use_4bit: bool = False):
     if not use_4bit:
         model = model.to(device)
 
+    # 确保自定义层使用正确的dtype
+    model.dist_proj = model.dist_proj.to(dtype=model_dtype)
+    model.dist_head = model.dist_head.to(dtype=model_dtype)
+
     model.eval()
     return model, tokenizer, config
 
@@ -206,7 +210,8 @@ def predict_distance_matrix(model, tokenizer, sequence: str, device: str = 'cuda
     with torch.no_grad():
         pred_dist = model(input_ids, attention_mask, num_of_pl_tokens=seq_len)
 
-    return pred_dist[0].cpu().numpy()
+    # 转换为float32再转numpy (bfloat16不支持numpy)
+    return pred_dist[0].float().cpu().numpy()
 
 
 def main():
